+++ test/TestAssemblies/PatternUsage/Attributes/Methods/DoubleOrIntVoTaskReturnAttribute.cs
@@ -1,0 +1,7 @@
+﻿namespace PatternUsage.Attributes.Methods
+{
+    public class DoubleOrIntVoTaskReturnAttribute : SetOnEntryAttribute
+    {
+        public override string? Pattern => "method(async double||int *(..))";
+    }
+}