+++ test/TestAssemblies/PatternUsage/Attributes/Methods/NoGenericVoTaskReturnAttribute.cs
@@ -1,0 +1,7 @@
+﻿namespace PatternUsage.Attributes.Methods
+{
+    public class NoGenericVoTaskReturnAttribute : SetOnEntryAttribute
+    {
+        public override string? Pattern => "method(async null *(..))";
+    }
+}